requirements.txt
altgraph==0.17.2
attrs==21.2.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
et-xmlfile==1.1.0
future==0.18.2
idna==3.3
isodate==0.6.0
lxml==4.6.3
numpy==1.21.2
openpyxl==3.0.8
pandas==1.3.2
peewee==3.14.4
pefile==2021.9.3
platformdirs==2.4.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
suds-community==1.0.0
tqdm==4.62.3
urllib3==1.26.7
zeep==4.1.0
